generate diagram on gh action and locally
On push, a github action will generate png diagrams and commit them.
Bot that commit on the repo are not ideal but we need a png version to
add it in the readme. With gh action generation we are sure the file
is up to date and we don't ask dev to generate it.

But if you need to generate it locally, you can use
`generate_diagram.sh` that will download and use the plantuml jar. See
$PLANTUML_JAR env var